+++ packages/web/src/modules/calendar-plan/components/inputs/CalendarPlanSelectionInput.vue
@@ -1,7 +1,14 @@
 <script lang="ts" setup>
-import { ISelectionDataPointValue, ISelectionDataPointSettings } from '@lyvely/common';
+import {
+  DataPointInputType,
+  ISelectionDataPointSettings,
+  ISelectionDataPointValue,
+} from '@lyvely/common';
 import { computed } from 'vue';
 import LyInputCheckbox from '@/modules/ui/components/form/CheckboxInput.vue';
+import SelectInput from '@/modules/ui/components/form/SelectInput.vue';
+import LyInputSelect from '@/modules/ui/components/form/SelectInput.vue';
+import LyInputRadio from '@/modules/ui/components/form/RadioInput.vue';
 
 interface IProps {
   modelValue: ISelectionDataPointValue;
@@ -24,14 +31,50 @@ const selection = computed({
     emit('update:modelValue', props.modelValue);
   },
 });
+
+const singleValueSelection = computed({
+  get: () => (props.modelValue.selection.length ? props.modelValue.selection[0] : undefined),
+
+  set: (selection: string) => {
+    // eslint-disable-next-line vue/no-mutating-props
+    props.modelValue.selection = [selection];
+    emit('update:modelValue', props.modelValue);
+  },
+});
+
+const dropDownOptions = computed(() =>
+  props.config.options.map((value) => ({ label: value, value })),
+);
 </script>
 
 <template>
-  <div class="flex flex-col gap-2 ml-2">
+  <div v-if="config.inputType === DataPointInputType.Checkbox" class="flex flex-col gap-2 ml-2">
     <div v-for="option in config.options" :key="option">
       <ly-input-checkbox v-model="selection" :label="option" :value="option" :translate="false" />
     </div>
   </div>
+  <div
+    v-else-if="config.inputType === DataPointInputType.Dropdown"
+    class="flex w-full items-center gap-2">
+    <ly-input-select
+      v-model="singleValueSelection"
+      class="calendar-plan-dropdown"
+      :options="dropDownOptions" />
+  </div>
+  <div
+    v-for="option in config.options"
+    v-else-if="config.inputType === DataPointInputType.Radio"
+    :key="option">
+    <ly-input-radio
+      v-model="singleValueSelection"
+      :label="option"
+      :value="option"
+      :translate="false" />
+  </div>
 </template>
 
-<style></style>
+<style scoped>
+.calendar-plan-dropdown {
+  min-width: 33.333333%;
+}
+</style>